Application: update BankAccountDTO, add Exceptions with normal messages in CQRS methods

+++ OnlineShop.Application/CQRS/BankAccounts/Commands/RemoveBankAccountById/RemoveBankAccountByIdCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.VisualBasic;
 using OnlineShop.Application.CQRS.BankAccounts.Handlers;
 using OnlineShop.Application.Repositories.Interfaces;
@@ -18,6 +19,13 @@ namespace OnlineShop.Application.CQRS.BankAccounts.Commands.RemoveBankAccountByI
 
         public async Task Handle(RemoveBankAccountByIdCommand request, CancellationToken cancellationToken)
         {
+            var bankAccount = await _repository.GetQuery().FirstOrDefaultAsync(x => x.Id == request.Id);
+
+            if (bankAccount is null)
+            {
+                throw new ArgumentException($"Not found BankAccount with id ({request.Id})");
+            }
+
             await _repository.RemoveByIdAsync(request.Id);
         }
     }
